Move PageWrapper to src/layouts. #86
